% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.r
\name{igraphFromSummary}
\alias{igraphFromSummary}
\title{Function to construct an igraph network from fold-change and p-value matrices}
\usage{
igraphFromSummary(fcMatrix, pvMatrix, cutoff, type = "removal")
}
\arguments{
\item{fcMatrix}{A matrix of fold changes where the rows are labelled with the strains affected and the columns with the strains either added or removed from the core experiment.}

\item{pvMatrix}{A matrix of p-values matching "fcMatrix".}

\item{cutoff}{A numeric indicating the p-value significance cutoff.}

\item{type}{A string, either "removal" or "addition", indicating the sort of experiment being summarised.}
}
\value{
An igraph network object.
}
\description{
Function to construct an igraph network from fold-change and p-value matrices
}
\details{
None
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
