% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{plotCommunityChanges}
\alias{plotCommunityChanges}
\title{Function to plot the changes in the community at a strain level.}
\usage{
plotCommunityChanges(cds, soi = NULL, cutoff = 0.05, rowLabs = NULL,
  subtitle = NULL, cols = 1:2, nBars = NULL, nLowPoints = TRUE)
}
\arguments{
\item{cds}{A compatible data set produced by the makeCDS() function.}

\item{soi}{A vector containing the names of the strains of interest, matching row names of the CDS count table.}

\item{cutoff}{A p-value cutoff to determine which strains will be included individually in the plot.}

\item{rowLabs}{A vector of names for the strains of interest, for plotting.}

\item{subtitle}{A subtitle for the plot.}

\item{cols}{A vector of two colours for the bars}

\item{nBars}{A numeric fixing the width of the plot to accommodate a certain number of bars}

\item{nLowPoints}{A logical indicating whether or not plot individual points instead of a box-and-whisker when the number of points is less than 4}
}
\value{
None.
}
\description{
Function to plot the changes in the community at a strain level.
}
\details{
None.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
