% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCommunityPie.r
\name{plotCommunityPie}
\alias{plotCommunityPie}
\title{Function to produce a pie chart of an individual community.}
\usage{
plotCommunityPie(counts, level = "Phlass", cols = phlassColors)
}
\arguments{
\item{counts}{A count table in which the rows are strains or OTUs and the columns are samples. The table should include only a single sample group.}

\item{level}{The taxonomic level at which the data will be summarised}

\item{cols}{A vector of colours where the names() of the vector correspond to the taxonomic labels.}
}
\value{
None.
}
\description{
Function to produce a pie chart of an individual community.
}
\details{
For each taxon at the chosen level, the sum of all counts for the strains in that taxon across all samples is divided by the total sum of the count table.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
