% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.r
\name{plotPCA}
\alias{plotPCA}
\title{Function to calculate and plot a PCA comparing sample groups at the whole community level.}
\usage{
plotPCA(cds, soi = NULL, perm = 100, cutoff = 0.05, rowLabs = NULL,
  subtitle = NULL, cols = 1:2, showLegend = TRUE,
  showArrows = TRUE, showHull = FALSE, showSidebars = FALSE,
  showTitle = TRUE, calcSpread = TRUE)
}
\arguments{
\item{cds}{A compatible data set produced by the makeCDS() function.}

\item{soi}{A vector containing the names of the strains of interest, matching row names of the CDS count table.}

\item{perm}{The number of permutations for the PERMANOVA, passed to the adonis() function.}

\item{cutoff}{A p-value cutoff to determine which strains will be included individually in the plot.}

\item{rowLabs}{A vector of names for the strains of interest, for plotting.}

\item{subtitle}{A subtitle for the plot.}

\item{cols}{A vector of 2 colours for the points.}

\item{showLegend}{A logical indicating whether or not to show a legend.}

\item{showArrows}{A logical indicating whether or not to show the top 3 or fewer significant individual responses}

\item{showHull}{A logical indicating whether or not to highlight the hull of each set of points}

\item{showSidebars}{A logical indicating whether or not to add box-and-whisker plots per group beside the axes}

\item{showTitle}{A logical indicating whether or not to show a plot title.}

\item{calcSpread}{A logical indicating whether or not to calculate the spread of each group of points}
}
\value{
The results of the PCA and the adonis() function, and optionally the spread of points in each group.
}
\description{
Function to calculate and plot a PCA comparing sample groups at the whole community level.
}
\details{
The spread of points is given as a vector of distances from each point in the group to that group's centroid.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
