% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseResults.r
\name{summariseResults}
\alias{summariseResults}
\title{Function to summarise the results of multiple phyllosphere experiments}
\usage{
summariseResults(cdsList)
}
\arguments{
\item{cdsList}{A list object containing compatible data sets created by "makeCDS". The names of each item in the list will be used to label the summary tables. The results in the data sets must contain the same rows in the same order, i.e.: this function is designed to summarise peturbations to a core experiment.}
}
\value{
A list containing two data frames: fcMatrix is the summary of log2FoldChange results, pvMatrix is the summary of adjusted p-value results (padj).
}
\description{
Function to summarise the results of multiple phyllosphere experiments
}
\details{
None
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
