% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.r
\name{treatmap}
\alias{treatmap}
\title{Function to plot a heatmap aligned with a phylogenetic tree.}
\usage{
treatmap(phylo, mat, mask = NULL, mask.color = "lightgrey",
  overlay = NULL, aspect.ratio = 1/exp(1), tip.labels = NULL,
  tip.colors = NULL, z.cols = NULL, z.res = NULL,
  tip.label.width = NULL, scalebar = T, mat.labels = NULL,
  mat.label.height = NULL, mat.label.colors = "black",
  mat.col.order = NULL, mat.hclust = FALSE, mat.phylo = NULL, ...)
}
\arguments{
\item{phylo}{An object of class "phylo".}

\item{mat}{A matrix of data for the heatmap with as many columns as tips in the phylogenetic tree.}

\item{mask}{An optional matrix of logicals equal in dimension to "mat" indicating whether some cells should not have their values plotted in the heatmap.}

\item{mask.color}{A color for the masked cells.}

\item{overlay}{An optional matrix of character strings to display on the cells of the heatmap, for instance significance labels.}

\item{aspect.ratio}{The aspect ratio width/height in which to plot the phylogenetic tree.}

\item{tip.labels}{A vector of character strings for tip labels, defaulting to those of the "phylo" object if not provided.}

\item{tip.colors}{A single color or vector of colors for the tip labels.}

\item{z.cols}{A vector of colors for the z-axis of the heatmap, see details for the default.}

\item{z.res}{A resolution for the color vector of the z-axis if automatically calculated.}

\item{tip.label.width}{A numeric indicating how much space to put between the heatmap and tree for plotting the labels, defaulting to 20% of the plotted tree width.}

\item{scalebar}{A logical indicating whether or not to include a scalebar in the plot.}

\item{mat.labels}{A vector of strings to label the columns of the heatmap; defaults to the column names of "mat".}

\item{mat.label.height}{A numeric indicating how much space to leave above the heatmap for column labels.}

\item{mat.label.colors}{A single color or vector of colors for the column labels of the heatmap.}

\item{mat.col.order}{A numeric vector indicating the order in which to plot the columns of the heatmap; ignored if mat.phylo is provided.}

\item{mat.hclust}{A logical indicating whether or not to hierarchically cluster the columns of the heatmap; ignored if mat.col.order or mat.phylo is provided.}

\item{mat.phylo}{An object of class "phylo" which is used to order the columns of the heatmap.}
}
\value{
None
}
\description{
Function to plot a heatmap aligned with a phylogenetic tree.
}
\details{
The default color vector for the z-axis is generated automatically if not provided. Firstly it is determined whether "mat" contains both positive and negative values. If it is one-sided then a vector of 32 colors is made with "colorRampPalette", between white and blue. If it is two-sided, a vector of 31 colors is made between blue, white and red. The argument "z.res" can be used to fix the number of colors in the vector, though it will still generate an odd number if the heatmap is two-sided. Providing "z.cols" will ignore all these determinations and split the color vector evenly between the heatmap's minimum and maximum.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
