% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylloR.r
\docType{package}
\name{phylloR}
\alias{phylloR}
\title{Package for microbiome analysis of the phyllosphere}
\description{
The package \pkg{phylloR} provides functions for analysing controlled experiments (not wild) involving the phyllosphere microbiome
}
\author{
Chris Field <fieldc@ethz.ch">
}
