% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\docType{package}
\name{phylloR}
\alias{phylloR}
\alias{makeCDS}
\alias{phylloR-package}
\title{Package for microbiome analysis of the phyllosphere}
\usage{
makeCDS(counts, meta, include = NULL, exclude = NULL, foi,
  ftc = NULL, title = NULL, legend = NULL)
}
\arguments{
\item{counts}{A count table in which the rows are strains or OTUs and the columns are samples.}

\item{meta}{A metadata table in which the rows correspond to the columns of the count table.}

\item{include}{A list that indicates which samples should be included in the analysis, see details.}

\item{exclude}{A list that indicates which samples should be excluded from the analysis, see details.}

\item{foi}{The factor of interest that determines how samples will be grouped for comparison.}

\item{ftc}{The factor to consider separately from the feature of interest, for instance to account for a batch effect.}

\item{title}{A title for the data set that will carry through to plots.}

\item{legend}{A vector with the names of the sample groups.}
}
\value{
A list containing the original dataset, the DESeq2-analysed dataset and the accompanying results.
}
\description{
The package \pkg{phylloR} provides functions for analysing controlled experiments (not wild) involving the phyllosphere microbiome
}
\details{
The lists for the include and exclude arguments should consist of vectors named to match the columns of the metadata being considered and contain the valid (or invalid) items.
For instance: include=list(initial_condition=c("Control","Test"),time=c("t1","t2")).
Samples are filtered first by the include list, then the exclude list. You can of course slice the data set in advance according to more complex criteria.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch">
Function to make a compatible data set (CDS) and calculate DESeq2 results.

Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
