% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_functions.r
\name{plotBipartiteSummary}
\alias{plotBipartiteSummary}
\title{Function to plot a bipartite graph to summarise strain-strain interactions.}
\usage{
plotBipartiteSummary(fcMatrix, pvMatrix, phylo = NULL,
  strainOrder = NULL, experiment.type = "removal",
  tip.label.width = 0, ...)
}
\arguments{
\item{fcMatrix}{A matrix of fold changes where the rows are labelled with the strains affected and the columns with the strains either added or removed from the core experiment.}

\item{pvMatrix}{A matrix of p-values matching "fcMatrix".}

\item{phylo}{An optional phylo object describing the phylogeny of the strains.}

\item{strainOrder}{An optional vector of character strings to place the strains in a particular order.}

\item{experiment.type}{A character string that is either "removal" or "addition" to describe the nature of the perturbations to the core experiment.}

\item{tip.label.width}{A numeric to indicate how much space should be placed between the tips of the tree (if provided) and graph lines.}

\item{...}{Other arguments passed to "draw.phylo" from the "apextra" package.}
}
\value{
None
}
\description{
Function to plot a bipartite graph to summarise strain-strain interactions.
}
\details{
The order in which the strains appear is determined by "strainOrder" preferentially, otherwise the order of the tips in "phylo", and otherwise as they are given in "fcMatrix".
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{None}
