% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{plotCommunity}
\alias{plotCommunity}
\title{Function to produce a bar or violin plot of an individual community.}
\usage{
plotCommunity(
  counts,
  type = "bar",
  xlabels = NULL,
  xcols = NULL,
  res = 50,
  cutoff = NULL
)
}
\arguments{
\item{counts}{A count table in which the rows are strains or OTUs and the columns are samples. The table should include only a single sample group.}

\item{type}{Determines the plot style, either 'bar', 'violin', 'points', 'swarm', 'barswarm' or 'violinswarm'.}

\item{xlabels}{An optional vector of strain names, the default is to use the row names of the count table.}

\item{xcols}{An optional vector of strain colours, the default is to use rainbow().}

\item{res}{The resolution of the histograms used to create the violins for that plot style.}

\item{cutoff}{A numeric integer, which if provided reduces the plot to only the given number of most abundant strains.}
}
\value{
A list of the statistics for each strain as generated by boxplot().
}
\description{
Function to produce a bar or violin plot of an individual community.
}
\details{
For each strain, samples in which they were not observed are counted and separated from the main bar or violin.
As a result, the number of samples in each bar or violin varies, and the histogram for each violin is therefore normalised.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
