% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{plotCommunityPie}
\alias{plotCommunityPie}
\title{Function to produce a pie chart of an individual community.}
\usage{
plotCommunityPie(counts, strainTaxa, cols, taxLabels = NULL,
  sort.tax = FALSE, drop = TRUE, ...)
}
\arguments{
\item{counts}{A count table in which the rows are strains or OTUs and the columns are samples. The table should include only a single sample group.}

\item{strainTaxa}{A vector of the taxonomic labels of the strains, in the order they appear in the rows of the count table.}

\item{cols}{A vector of colours where the names() of the vector correspond to the taxonomic labels.}

\item{taxLabels}{An optional vector defining the taxonomic labels to plot, with the default being the unique values of strainTaxa.}

\item{sort.tax}{If true, the taxonomic labels will be sorted from smallest to largest abundance.}

\item{drop}{A logical indicating whether unseen taxa should be dropped from the plot, defaults to TRUE.}

\item{...}{Further arguments to be passed to "pie".}
}
\value{
None.
}
\description{
Function to produce a pie chart of an individual community.
}
\details{
For each taxonomic label, the sum of all counts for the strains with that label across all samples is divided by the total sum of the count table.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
