% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCommunityStack.r
\name{plotCommunityStack}
\alias{plotCommunityStack}
\title{Function to produce a stacked bar chart of a phyllosphere community}
\usage{
plotCommunityStack(counts, level = "Phlass", meta = NULL,
  cols = phlassColors)
}
\arguments{
\item{counts}{A count table in which the rows are strains or OTUs and the columns are samples.}

\item{level}{Taxonomic level at which to summarise the community}

\item{meta}{Optional metadata that should be used to group samples}

\item{cols}{A named vector of colors for the different taxonomic levels}
}
\value{
None.
}
\description{
Function to produce a stacked bar chart of a phyllosphere community
}
\details{

}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
