% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.r
\name{plotPCA}
\alias{plotPCA}
\title{Function to calculate and plot a PCA comparing sample groups at the whole community level.}
\usage{
plotPCA(
  cds,
  soi = NULL,
  perm = 100,
  cutoff = 0.05,
  rowLabs = NULL,
  subtitle = NULL,
  cols = 1:2,
  showLegend = TRUE,
  showArrows = TRUE,
  showTitle = TRUE
)
}
\arguments{
\item{cds}{A compatible data set produced by the makeCDS() function.}

\item{soi}{A vector containing the names of the strains of interest, matching row names of the CDS count table.}

\item{perm}{The number of permutations for the PERMANOVA, passed to the adonis() function.}

\item{cutoff}{A p-value cutoff to determine which strains will be included individually in the plot.}

\item{rowLabs}{A vector of names for the strains of interest, for plotting.}

\item{subtitle}{A subtitle for the plot.}

\item{cols}{A vector of 2 colours for the points.}

\item{showLegend}{A logical indicating whether or not to show a legend.}

\item{showArrows}{A logical indicating whether or not to show the top 3 or fewer significant individual responses}

\item{showTitle}{A logical indicating whether or not to show a plot title.}
}
\value{
The results of the PCA and the adonis() function.
}
\description{
Function to calculate and plot a PCA comparing sample groups at the whole community level.
}
\details{
None.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
