% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.r
\name{plotVolcano}
\alias{plotVolcano}
\title{Function to produce a volcano plot of DESeq2 results}
\usage{
plotVolcano(cds, cutoff = 0.05, hili = TRUE, label = TRUE)
}
\arguments{
\item{cds}{A compatible data set produced by the makeCDS() function.}

\item{cutoff}{A p-value cutoff to determine which strains will be included individually in the plot.}

\item{hili}{Logical; whether to highlight significant points or not}

\item{label}{Logical; whether to label significant points or not}
}
\description{
Function to produce a volcano plot of DESeq2 results
}
\details{
None.
}
\examples{
None
}
\author{
Chris Field <fieldc@ethz.ch>
}
\keyword{phylloR}
